del by user without alert

+++ routes/feeds.js
@@ -2,7 +2,7 @@
 const route=require('express').Router()
 const Feeds=require('../db').Feeds
 route.get('/',(req,res)=>{
-    console.log("getting feeds")
+    //console.log("getting feeds")
     Feeds.findAll()
     .then((feeds)=>{
         res.status(200).send(feeds)
@@ -14,30 +14,39 @@ route.get('/',(req,res)=>{
 
 })
 route.post('/',(req,res)=>{
-    console.log("stat db")
-    console.log(req.body.author)
+   // console.log("stat db")
+  //  console.log(req.body.author)
     Feeds.create({
     
         author:req.user.username,
         content:req.body.content,
         
     }).then((feed)=>{
-        console.log("created")
+       // console.log("created")
         res.status(201).send(feed)
     }).catch((error)=>{
         res.status(501).send({
             error:"error adding feeds"
         })
     })
-    console.log("end db")
+   // console.log("end db")
 })
-route.post('/:id',(req,res)=>{
-    console.log(req.params.id)
-    Feeds.destroy({
+route.post('/:id/:author',(req,res)=>{
+     console.log(req.user.username)
+       console.log(req.params.author)
+    if(req.user.username===req.params.author){
+      //  console.log(req.user.username)
+       // console.log(req.params.author)
+       Feeds.destroy({
         where :{
             id:req.params.id
         }
     })
+    }else{
+        console.log("ggggggggggggggggggggggg")
+    }
+   // console.log(req.params.id)
+    
 })
